refactor: move telemetry classes to root package
