+++ openfeature-provider/src/test/java/com/spotify/confidence/FeatureProviderTest.java
@@ -14,8 +14,6 @@ import com.spotify.confidence.shaded.flags.resolver.v1.ResolveFlagsRequest;
 import com.spotify.confidence.shaded.flags.resolver.v1.ResolveFlagsResponse;
 import com.spotify.confidence.shaded.flags.resolver.v1.ResolvedFlag;
 import com.spotify.confidence.shaded.flags.types.v1.FlagSchema;
-import com.spotify.confidence.telemetry.Telemetry;
-import com.spotify.confidence.telemetry.TelemetryClientInterceptor;
 import com.spotify.telemetry.v1.LibraryTraces;
 import com.spotify.telemetry.v1.Monitoring;
 import dev.openfeature.sdk.*;
