+++ sdk-java/src/main/java/com/spotify/confidence/Confidence.java
@@ -16,8 +16,6 @@ import com.spotify.confidence.Exceptions.IncompatibleValueType;
 import com.spotify.confidence.Exceptions.ValueNotFound;
 import com.spotify.confidence.shaded.flags.resolver.v1.ResolveFlagsResponse;
 import com.spotify.confidence.shaded.flags.resolver.v1.ResolvedFlag;
-import com.spotify.confidence.telemetry.Telemetry;
-import com.spotify.confidence.telemetry.TelemetryClientInterceptor;
 import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
 import io.grpc.StatusRuntimeException;
