+++ sdk-java/src/main/java/com/spotify/confidence/GrpcFlagResolver.java
@@ -4,7 +4,6 @@ import com.google.common.base.Strings;
 import com.google.protobuf.Struct;
 import com.spotify.confidence.shaded.flags.resolver.v1.*;
 import com.spotify.confidence.shaded.flags.resolver.v1.Sdk.Builder;
-import com.spotify.confidence.telemetry.TelemetryClientInterceptor;
 import io.grpc.ManagedChannel;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
